+++ src/index.ts
@@ -1,13 +1,11 @@
 
-export {authorize, ready} from "./Client";
-
 import * as clienttypes from './Client';
 
 export namespace myTypes {
     export import clientypes = clienttypes;
 }
     //export function client(stateOrURI: fhirclient.ClientState | string): Client;
-    /*export const AbortController: {
-        new (): AbortController;
-        prototype: AbortController;
-    };*/
+/*export const AbortController: {
+    new (): AbortController;
+    prototype: AbortController;
+};*/
