+++ src/lib/absolute.ts
@@ -1,0 +1,14 @@
+/**
+ * Given a path, converts it to absolute url based on the `baseUrl`. If baseUrl
+ * is not provided, the result would be a rooted path (one that starts with `/`).
+ * @param path The path to convert
+ * @param baseUrl The base URL
+ */
+
+export function absolute(path: string, baseUrl?: string): string {
+    if (path.match(/^http/))
+        return path;
+    if (path.match(/^urn/))
+        return path;
+    return String(baseUrl || "").replace(/\/+$/, "") + "/" + path.replace(/^\/+/, "");
+}
