+++ src/lib/lib.ts
@@ -1,0 +1,19 @@
+/*
+ * This file contains some shared functions. They are used by other modules, but
+ * are defined here so that tests can import this library and test them.
+ */
+
+// $lab:coverage:off$
+// @ts-ignore
+//const { fetch } = typeof FHIRCLIENT_PURE !== "undefined" ? window : require("cross-fetch");
+// $lab:coverage:on$
+
+//const _debug     = debug("FHIR");
+//export { _debug as debug };
+
+/**
+ * The cache for the `getAndCache` function
+ */
+//const cache: Record<string, any> = {};
+
+
