+++ src/models/FetchOptions.ts
@@ -1,7 +1,7 @@
 /**
      * Options passed to the lib.request function
      */
- interface FetchOptions extends RequestInit {
+interface FetchOptions extends RequestInit {
     /**
      * If `true` the request function will be instructed to resolve with a
      * [[CombinedFetchResult]] object that contains the `Response` object
