+++ src/models/WellKnownSmartConfiguration.ts
