+++ src/types.ts
@@ -1,11 +1,11 @@
 export declare namespace openehrclient {
-    export function authorize(options: AuthorizeParams): Promise<string|void>;
+    export function authorize(options: AuthorizeParams): Promise<string | void>;
 
 
 
 
-type WindowTargetVariable = "_self"|"_top"|"_parent"|"_blank"|"popup"|string|number|Window;
-//    function WindowTargetFunction(): WindowTargetVariable;
-//    function WindowTargetFunction(): Promise<WindowTargetVariable>;
+    type WindowTargetVariable = "_self" | "_top" | "_parent" | "_blank" | "popup" | string | number | Window;
+    //    function WindowTargetFunction(): WindowTargetVariable;
+    //    function WindowTargetFunction(): Promise<WindowTargetVariable>;
     type WindowTarget = WindowTargetVariable //| typeof WindowTargetFunction;
 }
